requirements.txt
crewai
langchain
langchain-openai
openai
pandas
duckdb
plotly
yfinance
fastapi
uvicorn
python-dotenv
streamlit

